requirements.txt
--index-url https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cu126

torch==2.7.1+cu126
torchvision==0.22.1+cu126
torchaudio==2.7.1+cu126

numpy==1.26.4
scipy==1.11.4
pandas==2.2.3
tqdm==4.66.5

PyYAML==6.0.1
omegaconf==2.3.0
hydra-core==1.3.2
rich==13.7.1
einops==0.7.0

pillow==10.4.0
opencv-python==4.10.0.84
soundfile==0.12.1
librosa==0.10.2.post1
ffmpeg-python==0.2.0

matplotlib==3.8.4
requests==2.32.3
beautifulsoup4==4.12.3